+++ packages/components/inputs/time-input/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './time-input';
+
+export { default as version } from './version';
