--- src/components/inputs/time-input/index.js
@@ -1,1 +1,0 @@
-export { default } from './time-input';
